Add files via upload
In this Unity project, collision detection is implemented between a character (snowman named "Purly") and balloons, as well as between a moving sphere (snowball) and the character. When the character touches the balloons, they get destroyed, while the character is destroyed upon collision with the snowball. The snowball follows the character's movements and detects collisions using trigger colliders and custom scripting.